2 PRs approach two fix diff
